fix a cast (bug 694546)

+++ src/mesa/swrast/s_aaline.c
@@ -1,10 +1,10 @@
-/* $Id: s_aaline.c,v 1.16 2002/08/07 00:45:07 brianp Exp $ */
+/* $Id: s_aaline.c,v 1.16.4.1 2003/02/27 23:38:23 brianp Exp $ */
 
 /*
  * Mesa 3-D graphics library
- * Version:  4.1
+ * Version:  5.0.1
  *
- * Copyright (C) 1999-2002  Brian Paul   All Rights Reserved.
+ * Copyright (C) 1999-2003  Brian Paul   All Rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -184,7 +184,7 @@ solve_plane_chan(GLfloat x, GLfloat y, const GLfloat plane[4])
       return 0;
    else if (z > CHAN_MAXF)
       return (GLchan) CHAN_MAXF;
-   return (GLchan) (GLint) z;
+   return (GLchan) z;
 }
